Move client internals back to Client class. Remove obsolete ClientBase class
